Add Browser.application test (not fully supported yet)
HACK replace the Browser.Navigation.Key in the model (needs tag from Evan!)

I'm hardcoding the location in my example App's model for now. But
we will need a real solution soon.

One way to make it work is adding this tag to the navigation key in
`Browser_application`

    key["$HMR$"] = "Browser.Navigation.Key";

Right before the line:

    return A3(impl.init, flags, _Browser_getUrl(), key);