Big change: track messages instead of model state
Evan recommended that rather than storing the last
model value, instead store the messages which have
been delivered to the app and then replay them
after the new code is loaded. This is how the Elm
time-travelling debugger works, and we're kinda
forced to do it now that `Browser.application` has
this `Key` value which is stored in the user's model
but is actually a function reference to Elm's
dispatch / effect-manager system.