Find (and restore) Browser.Navigation.Key values in the model
If you don't do this, then Browser.application will be broken
after HMR because the hot swap will preserve the old key. But
the old key is only valid for the old app. You must use the
new key in the new app, otherwise pushUrl (and friends) will
be broken.